build dockerfile taotao-cloud-gateway taotao-cloud-monitor taotao-cloud-file  taotao-cloud-job  taotao-cloud-auth taotao-cloud-sys  taotao-cloud-xxljob

+++ taotao-cloud-microservice/taotao-cloud-business/taotao-cloud-payment/taotao-cloud-payment-biz/src/main/java/com/taotao/cloud/payment/biz/jeepay/service/impl/MchInfoService.java
@@ -1,46 +1,42 @@
 /*
- * Copyright (c) 2020-2030, Shuigedeng (981376577@qq.com & https://blog.taotaocloud.top/).
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
+ * Copyright (c) 2021-2031, 河北计全科技有限公司 (https://www.jeequan.com & jeequan@126.com).
+ * <p>
+ * Licensed under the GNU LESSER GENERAL PUBLIC LICENSE 3.0;
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p>
+ * http://www.gnu.org/licenses/lgpl.html
+ * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package com.taotao.cloud.payment.biz.jeepay.service.impl;
 
-import org.dromara.hutoolcore.util.IdUtil;
-import org.dromara.hutoolcore.util.RandomUtil;
+import cn.hutool.core.util.IdUtil;
+import cn.hutool.core.util.RandomUtil;
+import com.baomidou.mybatisplus.core.toolkit.CollectionUtils;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
-import com.taotao.cloud.payment.biz.jeepay.core.constants.ApiCodeEnum;
-import com.taotao.cloud.payment.biz.jeepay.core.constants.CS;
-import com.taotao.cloud.payment.biz.jeepay.core.entity.IsvInfo;
-import com.taotao.cloud.payment.biz.jeepay.core.entity.MchApp;
-import com.taotao.cloud.payment.biz.jeepay.core.entity.MchInfo;
-import com.taotao.cloud.payment.biz.jeepay.core.entity.MchPayPassage;
-import com.taotao.cloud.payment.biz.jeepay.core.entity.PayInterfaceConfig;
-import com.taotao.cloud.payment.biz.jeepay.core.entity.PayOrder;
-import com.taotao.cloud.payment.biz.jeepay.core.entity.SysUser;
-import com.taotao.cloud.payment.biz.jeepay.core.entity.SysUserAuth;
-import com.taotao.cloud.payment.biz.jeepay.core.exception.BizException;
-import com.taotao.cloud.payment.biz.jeepay.service.mapper.MchInfoMapper;
-import java.util.ArrayList;
-import java.util.LinkedList;
-import java.util.List;
+import com.jeequan.jeepay.core.constants.ApiCodeEnum;
+import com.jeequan.jeepay.core.constants.CS;
+import com.jeequan.jeepay.core.entity.*;
+import com.jeequan.jeepay.core.exception.BizException;
+import com.jeequan.jeepay.service.mapper.MchInfoMapper;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
+import java.util.LinkedList;
+import java.util.List;
+
 /**
+ * <p>
  * 商户信息表 服务实现类
+ * </p>
  *
  * @author [mybatis plus generator]
  * @since 2021-04-27
@@ -48,26 +44,19 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 public class MchInfoService extends ServiceImpl<MchInfoMapper, MchInfo> {
 
-    @Autowired
-    private SysUserService sysUserService;
+    @Autowired private SysUserService sysUserService;
 
-    @Autowired
-    private PayOrderService payOrderService;
+    @Autowired private PayOrderService payOrderService;
 
-    @Autowired
-    private MchPayPassageService mchPayPassageService;
+    @Autowired private MchPayPassageService mchPayPassageService;
 
-    @Autowired
-    private PayInterfaceConfigService payInterfaceConfigService;
+    @Autowired private PayInterfaceConfigService payInterfaceConfigService;
 
-    @Autowired
-    private SysUserAuthService sysUserAuthService;
+    @Autowired private SysUserAuthService sysUserAuthService;
 
-    @Autowired
-    private IsvInfoService isvInfoService;
+    @Autowired private IsvInfoService isvInfoService;
 
-    @Autowired
-    private MchAppService mchAppService;
+    @Autowired private MchAppService mchAppService;
 
     @Transactional(rollbackFor = Exception.class)
     public void addMch(MchInfo mchInfo, String loginUserName) {
@@ -121,9 +110,10 @@ public class MchInfoService extends ServiceImpl<MchInfoMapper, MchInfo> {
         if (!saveResult) {
             throw new BizException(ApiCodeEnum.SYS_OPERATION_FAIL_CREATE);
         }
+
     }
 
-    /** 删除商户 * */
+    /** 删除商户 **/
     @Transactional(rollbackFor = Exception.class)
     public List<Long> removeByMchNo(String mchNo) {
         try {
@@ -134,7 +124,7 @@ public class MchInfoService extends ServiceImpl<MchInfoMapper, MchInfo> {
             }
 
             // 1.查看当前商户是否存在交易数据
-            long payCount = payOrderService.count(PayOrder.gw().eq(PayOrder::getMchNo, mchNo));
+            int payCount = payOrderService.count(PayOrder.gw().eq(PayOrder::getMchNo, mchNo));
             if (payCount > 0) {
                 throw new BizException("该商户已存在交易数据，不可删除");
             }
@@ -145,20 +135,27 @@ public class MchInfoService extends ServiceImpl<MchInfoMapper, MchInfo> {
             // 3.删除当前商户支付接口配置参数
             List<String> appIdList = new LinkedList<>();
             mchAppService.list(MchApp.gw().eq(MchApp::getMchNo, mchNo)).forEach(item -> appIdList.add(item.getAppId()));
-            payInterfaceConfigService.remove(PayInterfaceConfig.gw()
-                    .in(PayInterfaceConfig::getInfoId, appIdList)
-                    .eq(PayInterfaceConfig::getInfoType, CS.INFO_TYPE_MCH_APP));
+            if (CollectionUtils.isNotEmpty(appIdList)) {
+                payInterfaceConfigService.remove(PayInterfaceConfig.gw()
+                        .in(PayInterfaceConfig::getInfoId, appIdList)
+                        .eq(PayInterfaceConfig::getInfoType, CS.INFO_TYPE_MCH_APP)
+                );
+            }
 
-            List<SysUser> userList = sysUserService.list(
-                    SysUser.gw().eq(SysUser::getBelongInfoId, mchNo).eq(SysUser::getSysType, CS.SYS_TYPE.MCH));
+            List<SysUser> userList = sysUserService.list(SysUser.gw()
+                    .eq(SysUser::getBelongInfoId, mchNo)
+                    .eq(SysUser::getSysType, CS.SYS_TYPE.MCH)
+            );
 
             // 4.删除当前商户应用信息
-            mchAppService.removeByIds(appIdList);
+            if (CollectionUtils.isNotEmpty(appIdList)) {
+                mchAppService.removeByIds(appIdList);
+            }
 
             // 返回的用户id
             List<Long> userIdList = new ArrayList<>();
-            if (userList.size() > 0) {
-                for (SysUser user : userList) {
+            if (CollectionUtils.isNotEmpty(userList)) {
+                for (SysUser user:userList) {
                     userIdList.add(user.getSysUserId());
                 }
                 // 5.删除当前商户用户子用户信息
@@ -166,8 +163,10 @@ public class MchInfoService extends ServiceImpl<MchInfoMapper, MchInfo> {
             }
 
             // 6.删除当前商户的登录用户
-            sysUserService.remove(
-                    SysUser.gw().eq(SysUser::getBelongInfoId, mchNo).eq(SysUser::getSysType, CS.SYS_TYPE.MCH));
+            sysUserService.remove(SysUser.gw()
+                    .eq(SysUser::getBelongInfoId, mchNo)
+                    .eq(SysUser::getSysType, CS.SYS_TYPE.MCH)
+            );
 
             // 7.删除当前商户
             boolean removeMchInfo = removeById(mchNo);
@@ -175,7 +174,7 @@ public class MchInfoService extends ServiceImpl<MchInfoMapper, MchInfo> {
                 throw new BizException("删除当前商户失败");
             }
             return userIdList;
-        } catch (Exception e) {
+        }catch (Exception e) {
             throw new BizException(e.getMessage());
         }
     }
